Juniper: move parboiled classes into a specific bazel package
Code coverage can only refuse to instrument classes on a per-Bazel-package basis.

Indirect utility classes with simple public API (e.g., String->String) to keep
the public API the same.

commit-id:193e851f
