+++ projects/batfish/src/main/java/org/batfish/representation/juniper/parboiled/BUILD
@@ -1,0 +1,26 @@
+load("@rules_java//java:defs.bzl", "java_library")
+load("@batfish//skylark:pmd_test.bzl", "pmd_test")
+
+package(default_visibility = [
+    "//projects/batfish/src/main/java/org/batfish/representation/juniper:__subpackages__",
+])
+
+java_library(
+    name = "parboiled",
+    srcs = [
+        "AsPathRegex.java",
+        "GroupWildcard.java",
+    ],
+    deps = [
+        "//projects/batfish-common-protocol:common",
+        "@maven//:com_google_code_findbugs_jsr305",
+        "@maven//:com_google_guava_guava",
+        "@maven//:org_parboiled_parboiled_core",
+        "@maven//:org_parboiled_parboiled_java",
+    ],
+)
+
+pmd_test(
+    name = "parboiled_pmd",
+    lib = ":parboiled",
+)
